Consolodate findIncludePath
